Add User Auth

+++ app/Http/Controllers/TasksController.php
@@ -16,11 +16,27 @@ class TasksController extends Controller
      */
     public function index()
     {
-        //タスク一覧を取得
-        $tasks = Task::all();
+        // //タスク一覧を取得
+        // $tasks = Task::all();
         
-        // タスク一覧ビューで表示
-        return view('tasks.index', ['tasks' => $tasks,]);
+        // // タスク一覧ビューで表示
+        // return view('tasks.index', ['tasks' => $tasks,]);
+        $data = [];
+        if (\Auth::check()) {
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            
+            // ユーザのタスクリスト一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        
+        // トップビューで表示
+        return view('top', $data);
     }
 
     /**
@@ -53,10 +69,14 @@ class TasksController extends Controller
         ]);
         
         //タスクを作成
-        $task = New Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        // $task = New Task;
+        // $task->content = $request->content;
+        // $task->status = $request->status;
+        // $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
         
         // トップページへリダイレクトする
         return redirect('/');
@@ -74,8 +94,12 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         
-        // タスク詳細ビューで表示
-        return view('tasks.show', ['task' => $task,]);
+        if (\Auth::id() === $task->user_id) {
+            // タスク詳細ビューで表示
+            return view('tasks.show', ['task' => $task,]);
+        }else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -90,8 +114,12 @@ class TasksController extends Controller
         //idの値でタスクを検索して取得
         $task = Task::FindOrFail($id);
         
-        // タスク編集ビューで表示
-        return view('tasks.edit',['task' => $task,]);
+        if (\Auth::id() === $task->user_id) {
+            // タスク編集ビューで表示
+            return view('tasks.edit',['task' => $task,]);
+        }else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -135,9 +163,11 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::FindOrFail($id);
         
-        // 該当データがあった場合に
-        // タスクを削除
-        $task->delete();
+        if (\Auth::id() === $task->user_id) {
+            // 該当データがあった場合に
+            // タスクを削除
+            $task->delete();
+        }
         
         // トップページへリダイレクト
         return redirect('/');
